requirements.txt
Django==4.2.2
djangorestframework==3.14.0
drf-yasg==1.21.6
psycopg2-binary==2.9.6
psycopg2>=2.8
python-dotenv==1.0.0
pytest==7.4.0
pytest-django==4.5.2